Split jobs/handlers.ts and move OPML-import processing into a service (#1470)
handlers.ts was 1928 lines mixing the feed-fetch pipeline with six unrelated
singleton-job handlers. It becomes handlers/, one file per job type (plus the
shared JobHandlerResult in handlers/types.ts), so adding a job type means adding
a file rather than appending to an existing handler's.

The ~400 lines of import business logic in handleProcessOpmlImport (per-feed
loop, tag resolution, progress publishing, status updates) move into
services/imports.ts as processOpmlImport, next to the parse+enqueue half; the
handler is now scheduling glue that maps the service's outcome to a job result.
That processing was previously untested — integration tests now cover the import
loop, the already-subscribed skip, the crashed-run recovery path, and a missing
import record.

No behavior change otherwise.
