+++ docs/diagrams/sse-cache-updates.d2
@@ -45,7 +45,7 @@ event-sources: Event Sources {
 
   job-handlers: Background Jobs {
     style.fill: "#e0e7ff"
-    tooltip: "src/server/jobs/handlers.ts"
+    tooltip: "src/server/services/imports.ts"
 
     import-sub-created: publishSubscriptionCreated(...) // OPML import
     import-progress: publishImportProgress(userId, importId, feedUrl, status, counts)
