+++ src/server/jobs/handlers/fetch-feed.ts
@@ -1,87 +1,43 @@
 /**
- * Job handlers for processing different job types.
+ * Handler for `fetch_feed` jobs: the feed-fetch pipeline.
  *
- * Each handler executes a specific job type and returns a result that includes
- * the next run time for the job. The worker uses this to update the job record.
+ * Fetches a feed, applies the fetch result (parse + entry processing, backoff on
+ * error, permanent-redirect tracking/migration) and reports the next fetch time.
  *
- * See docs/job-queue-design.md for the overall architecture.
+ * See docs/job-queue-design.md for the overall architecture and
+ * docs/diagrams/feed-fetcher.d2 for the pipeline.
  */
 
 import { createHash } from "crypto";
 import { eq, and, isNull, inArray, count } from "drizzle-orm";
-import { db } from "../db";
+import { db } from "../../db";
+import { feeds, subscriptions, entries, userEntries, type Feed } from "../../db/schema";
+import { fetchFullContent, persistFullContentResult } from "../../services/full-content";
+import { fetchFeed, type FetchFeedResult, type RedirectInfo } from "../../feed/fetcher";
+import type { WebSubLinkHeaders } from "../../feed/link-header";
+import { parseFeed } from "../../feed/parser";
+import { processEntries } from "../../feed/entry-processor";
+import { calculateNextFetch } from "../../feed/scheduling";
 import {
-  feeds,
-  subscriptions,
-  entries,
-  opmlImports,
-  tags,
-  subscriptionTags,
-  userEntries,
-  type Feed,
-  type OpmlImportFeedResult,
-} from "../db/schema";
-import { fetchFullContent, persistFullContentResult } from "../services/full-content";
-import { fetchFeed, type FetchFeedResult, type RedirectInfo } from "../feed/fetcher";
-import type { WebSubLinkHeaders } from "../feed/link-header";
-import { parseFeed } from "../feed/parser";
-import { processEntries } from "../feed/entry-processor";
-import { calculateNextFetch } from "../feed/scheduling";
-import {
-  renewExpiringSubscriptions,
   canUseWebSub,
   subscribeToHub,
   deactivateWebsub,
   resolveWebsubAction,
-} from "../feed/websub";
-import { recordBackupPollNewEntries } from "../feed/websub-hub-stats";
-import { getDomainFromUrl } from "../feed/types";
-import type { ParsedCacheHeaders } from "../feed/cache-headers";
-import {
-  getFeedFetchHealthSnapshot,
-  evaluateFeedFetchHealth,
-  buildFeedHealthPingBody,
-} from "../feed/health";
-import { pingHealthcheck } from "../notifications/healthchecks";
-import { feedHealthConfig } from "../config/env";
-import { type JobPayloads, ensureFeedJob } from "./queue";
+} from "../../feed/websub";
+import { recordBackupPollNewEntries } from "../../feed/websub-hub-stats";
+import { getDomainFromUrl } from "../../feed/types";
+import type { ParsedCacheHeaders } from "../../feed/cache-headers";
+import { type JobPayloads, ensureFeedJob } from "../queue";
 import { logger } from "@/lib/logger";
-import {
-  startFeedFetchTimer,
-  trackWebsubRenewal,
-  updateFeedHealthMetrics,
-  type FeedFetchStatus,
-} from "../metrics/metrics";
-import {
-  publishImportProgress,
-  publishImportCompleted,
-  publishSubscriptionUpdated,
-} from "../redis/pubsub";
+import { startFeedFetchTimer, type FeedFetchStatus } from "../../metrics/metrics";
 import { generateUuidv7 } from "@/lib/uuidv7";
 import { getFeedPlugin } from "@/server/plugins";
-import { createSubscription } from "../services/subscriptions";
-import { runRetentionCleanup } from "../services/retention";
-import { reconcileCounters } from "../services/reconcile-counters";
-import { backfillGettingStartedArticles } from "../services/getting-started";
 import {
   findPermanentRedirectUrl,
   isHttpToHttpsUpgrade,
   REDIRECT_WAIT_PERIOD_MS,
-} from "../feed/redirect-utils";
-
-/**
- * Result of a job handler execution.
- */
-export interface JobHandlerResult {
-  /** Whether the job completed successfully */
-  success: boolean;
-  /** When the job should next run */
-  nextRunAt: Date;
-  /** Error message if the job failed */
-  error?: string;
-  /** Any additional metadata about the job execution */
-  metadata?: Record<string, unknown>;
-}
+} from "../../feed/redirect-utils";
+import type { JobHandlerResult } from "./types";
 
 /**
  * Maximum number of entries to fetch full content for per feed fetch.
@@ -1279,650 +1235,4 @@ export async function migrateSubscriptionsToExistingFeed(
 
   // Note: The old feed's job will naturally stop being claimed since it has no active
   // subscribers. In the data-driven model, we don't need to explicitly disable it.
-}
-
-/**
- * How often the WebSub renewal job runs.
- *
- * This must be short relative to the shortest lease we want to keep alive: a
- * lease can only be renewed on a run that falls within its renewal window, so a
- * daily cadence silently lets sub-24h leases lapse until the next run. Running
- * hourly keeps any lease down to ~1h alive.
- */
-const WEBSUB_RENEWAL_INTERVAL_MS = 60 * 60 * 1000; // 1 hour
-
-/**
- * How far ahead of expiry we renew, in hours.
- *
- * Kept just above the run interval so every lease is renewed on the run before
- * it would expire, while long leases are left untouched until near their own
- * expiry (no per-run re-subscribe spam).
- */
-const WEBSUB_RENEWAL_THRESHOLD_HOURS = 2;
-
-/**
- * Handler for renew_websub jobs.
- * Renews WebSub subscriptions that are expiring soon.
- *
- * Runs hourly (see WEBSUB_RENEWAL_INTERVAL_MS) and renews active subscriptions
- * expiring within WEBSUB_RENEWAL_THRESHOLD_HOURS. The frequent cadence + short
- * threshold keeps short leases alive without re-subscribing long leases every
- * run.
- *
- * @param _payload - The job payload (empty for this job type)
- * @returns Job handler result with next run time
- */
-export async function handleRenewWebsub(
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _payload: JobPayloads["renew_websub"]
-): Promise<JobHandlerResult> {
-  logger.info("Starting WebSub subscription renewal check");
-
-  const result = await renewExpiringSubscriptions(WEBSUB_RENEWAL_THRESHOLD_HOURS);
-
-  // Track renewal metrics
-  for (let i = 0; i < result.renewed; i++) {
-    trackWebsubRenewal(true);
-  }
-  for (let i = 0; i < result.failed; i++) {
-    trackWebsubRenewal(false);
-  }
-
-  if (result.failed > 0) {
-    logger.warn("Some WebSub renewals failed", {
-      errors: result.errors,
-    });
-  }
-
-  // Schedule the next check one interval out.
-  const nextRunAt = new Date(Date.now() + WEBSUB_RENEWAL_INTERVAL_MS);
-
-  logger.debug("Scheduled next WebSub renewal check", {
-    scheduledFor: nextRunAt.toISOString(),
-  });
-
-  return {
-    success: true,
-    nextRunAt,
-    metadata: {
-      checked: result.checked,
-      renewed: result.renewed,
-      failed: result.failed,
-      errors: result.errors.length > 0 ? result.errors : undefined,
-    },
-  };
-}
-
-/** How often the feed fetch health check runs. */
-const FEED_HEALTH_CHECK_INTERVAL_MS = 15 * 60 * 1000;
-
-/**
- * Handler for monitor_feed_health jobs (singleton, runs every 15 minutes).
- *
- * Checks the invariant "at least one feed fetched successfully recently"
- * (see src/server/feed/health.ts) and:
- * - Pings the configured healthchecks.io check (FEED_HEALTH_HEARTBEAT_URL):
- *   a success ping when healthy, a `/fail` ping with an explanatory body when
- *   not. This is the feed-fetch *quality* signal; the worker process's own
- *   liveness is a separate check (WORKER_HEARTBEAT_URL, see scripts/worker.ts),
- *   so "feeds are failing" stays distinguishable from "worker is dead".
- * - Updates the feed health Prometheus gauges.
- *
- * No alert state is kept here: healthchecks.io de-duplicates notifications and
- * sends its own recovery ("up") email, so the job just reports status each run.
- */
-export async function handleMonitorFeedHealth(
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _payload: JobPayloads["monitor_feed_health"]
-): Promise<JobHandlerResult> {
-  const now = new Date();
-  const snapshot = await getFeedFetchHealthSnapshot();
-  const evaluation = evaluateFeedFetchHealth(
-    snapshot,
-    now,
-    feedHealthConfig.maxSuccessAgeMinutes * 60 * 1000
-  );
-
-  updateFeedHealthMetrics(
-    evaluation.lastSuccessAgeMs !== null ? evaluation.lastSuccessAgeMs / 1000 : null,
-    snapshot.failingFeedCount
-  );
-
-  if (evaluation.status === "unhealthy") {
-    logger.warn("Feed fetch health check failed", {
-      reason: evaluation.reason,
-      lastSuccessfulFetchAt: snapshot.lastSuccessfulFetchAt?.toISOString(),
-      failingFeedCount: snapshot.failingFeedCount,
-      pollableFeedCount: snapshot.pollableFeedCount,
-      sampleError: snapshot.sampleError,
-    });
-  }
-
-  if (feedHealthConfig.heartbeatUrl) {
-    await pingHealthcheck(feedHealthConfig.heartbeatUrl, {
-      signal: evaluation.status === "healthy" ? "success" : "fail",
-      body: buildFeedHealthPingBody(snapshot, evaluation),
-    });
-  }
-
-  return {
-    success: true,
-    nextRunAt: new Date(now.getTime() + FEED_HEALTH_CHECK_INTERVAL_MS),
-    metadata: {
-      status: evaluation.status,
-      reason: evaluation.reason,
-      lastSuccessAgeMs: evaluation.lastSuccessAgeMs ?? undefined,
-      failingFeedCount: snapshot.failingFeedCount,
-      pollableFeedCount: snapshot.pollableFeedCount,
-    },
-  };
-}
-
-/** How often the retention cleanup runs. */
-const CLEANUP_INTERVAL_MS = 24 * 60 * 60 * 1000;
-
-/**
- * Handler for cleanup jobs (singleton, runs daily).
- *
- * Deletes rows that expire but were never deleted anywhere (issue #953):
- * expired sessions, expired OAuth authorization codes / access tokens /
- * refresh tokens, long-revoked credentials, orphaned Dynamic Client
- * Registration clients (issue #975), parked one-time process_opml_import
- * jobs, and subscriber-less fetch_feed jobs (issue #1085). See
- * src/server/services/retention.ts.
- */
-export async function handleCleanup(
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _payload: JobPayloads["cleanup"]
-): Promise<JobHandlerResult> {
-  const now = new Date();
-  const deleted = await runRetentionCleanup(db);
-
-  logger.info("Retention cleanup completed", { ...deleted });
-
-  return {
-    success: true,
-    nextRunAt: new Date(now.getTime() + CLEANUP_INTERVAL_MS),
-    metadata: { ...deleted },
-  };
-}
-
-/**
- * How often the denormalized unread counters are reconciled against ground
- * truth. Daily: the counter triggers should keep them exact, so this is a
- * drift detector (any fix logs at error level) that doubles as self-healing.
- */
-const RECONCILE_COUNTERS_INTERVAL_MS = 24 * 60 * 60 * 1000;
-
-/**
- * Handler for reconcile_counters jobs (singleton, stateless, runs daily).
- * See src/server/services/reconcile-counters.ts.
- */
-export async function handleReconcileCounters(
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _payload: JobPayloads["reconcile_counters"]
-): Promise<JobHandlerResult> {
-  const now = new Date();
-  const result = await reconcileCounters(db);
-
-  return {
-    success: true,
-    nextRunAt: new Date(now.getTime() + RECONCILE_COUNTERS_INTERVAL_MS),
-    metadata: { ...result },
-  };
-}
-
-/** Users given a Getting Started article per backfill run. */
-const GETTING_STARTED_BATCH_SIZE = 100;
-
-/** Gap between backfill runs while users are still waiting. */
-const GETTING_STARTED_BACKLOG_INTERVAL_MS = 15 * 1000;
-
-/**
- * Gap once the backlog is drained. The job stays alive rather than parking:
- * the same scan also catches anyone whose signup-path insert failed, and once
- * the partial index is empty the query costs nothing.
- */
-const GETTING_STARTED_IDLE_INTERVAL_MS = 24 * 60 * 60 * 1000;
-
-/**
- * Handler for backfill_getting_started jobs (singleton, stateless).
- * See src/server/services/getting-started.ts.
- */
-export async function handleBackfillGettingStarted(
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _payload: JobPayloads["backfill_getting_started"]
-): Promise<JobHandlerResult> {
-  const result = await backfillGettingStartedArticles(db, GETTING_STARTED_BATCH_SIZE);
-
-  // Come back promptly while we're filling batches and getting somewhere. A
-  // batch where *nothing* succeeded drops to the idle cadence rather than
-  // spinning every 15 seconds; a batch with a few stuck users still counts as
-  // progress, so one poison row can't throttle the whole backfill to daily.
-  const moreLikely = result.attempted === GETTING_STARTED_BATCH_SIZE && result.created > 0;
-
-  return {
-    success: true,
-    nextRunAt: new Date(
-      Date.now() +
-        (moreLikely ? GETTING_STARTED_BACKLOG_INTERVAL_MS : GETTING_STARTED_IDLE_INTERVAL_MS)
-    ),
-    metadata: { ...result },
-  };
-}
-
-// Batch size for database updates during OPML import.
-// Instead of updating the database after every feed (which causes freezes with large imports),
-// we batch updates to reduce database load. Redis events still fire per-feed for real-time UI.
-const OPML_IMPORT_DB_BATCH_SIZE = 10;
-
-/**
- * Handler for process_opml_import jobs.
- * Processes an OPML import in the background, publishing progress events
- * as each feed is processed.
- *
- * This is a one-time job - it completes and is not rescheduled.
- *
- * @param payload - The job payload containing the importId
- * @returns Job handler result
- */
-export async function handleProcessOpmlImport(
-  payload: JobPayloads["process_opml_import"]
-): Promise<JobHandlerResult> {
-  const { importId } = payload;
-
-  logger.info("Starting OPML import processing", { importId });
-
-  // Get the import record
-  const [importRecord] = await db
-    .select()
-    .from(opmlImports)
-    .where(eq(opmlImports.id, importId))
-    .limit(1);
-
-  if (!importRecord) {
-    logger.warn("Import record not found", { importId });
-    return {
-      success: false,
-      nextRunAt: new Date(Date.now() + 365 * 24 * 60 * 60 * 1000), // Far future - job won't run again
-      error: `Import record not found: ${importId}`,
-    };
-  }
-
-  // If already completed or failed, don't reprocess
-  if (importRecord.status === "completed" || importRecord.status === "failed") {
-    logger.info("Import already finished, skipping", { importId, status: importRecord.status });
-    return {
-      success: true,
-      nextRunAt: new Date(Date.now() + 365 * 24 * 60 * 60 * 1000), // Far future - job won't run again
-    };
-  }
-
-  const userId = importRecord.userId;
-
-  // Check if a previous run completed processing but crashed before updating status.
-  // This can happen if the worker was killed between finishing the loop and the final update.
-  // We detect this by checking if all feeds have been processed, using either:
-  // 1. The results array length matching total feeds (most accurate)
-  // 2. The counts adding up to total feeds (for cases where results array wasn't fully saved)
-  const processedByResults = importRecord.results.length === importRecord.totalFeeds;
-  const processedByCounts =
-    importRecord.importedCount + importRecord.skippedCount + importRecord.failedCount ===
-    importRecord.totalFeeds;
-
-  if (importRecord.status === "processing" && (processedByResults || processedByCounts)) {
-    // Use counts from results array if available, otherwise use stored counts
-    const imported = processedByResults
-      ? importRecord.results.filter((r) => r.status === "imported").length
-      : importRecord.importedCount;
-    const skipped = processedByResults
-      ? importRecord.results.filter((r) => r.status === "skipped").length
-      : importRecord.skippedCount;
-    const failed = processedByResults
-      ? importRecord.results.filter((r) => r.status === "failed").length
-      : importRecord.failedCount;
-
-    logger.info("Recovering previously completed import", {
-      importId,
-      imported,
-      skipped,
-      failed,
-      total: importRecord.totalFeeds,
-    });
-
-    // Mark import as completed
-    const completedAt = new Date();
-    await db
-      .update(opmlImports)
-      .set({
-        status: "completed",
-        importedCount: imported,
-        skippedCount: skipped,
-        failedCount: failed,
-        completedAt,
-        updatedAt: completedAt,
-      })
-      .where(eq(opmlImports.id, importId));
-
-    // Publish completed event
-    await publishImportCompleted(userId, importId, {
-      imported,
-      skipped,
-      failed,
-      total: importRecord.totalFeeds,
-    });
-
-    return {
-      success: true,
-      nextRunAt: new Date(Date.now() + 365 * 24 * 60 * 60 * 1000), // Far future - one-time job
-      metadata: {
-        recovered: true,
-        imported,
-        skipped,
-        failed,
-        total: importRecord.totalFeeds,
-      },
-    };
-  }
-
-  // Update status to processing
-  await db
-    .update(opmlImports)
-    .set({ status: "processing", updatedAt: new Date() })
-    .where(eq(opmlImports.id, importId));
-
-  try {
-    // Get existing subscriptions for the user
-    const existingSubscriptions = await db
-      .select({
-        feedUrl: feeds.url,
-      })
-      .from(subscriptions)
-      .innerJoin(feeds, eq(subscriptions.feedId, feeds.id))
-      .where(and(eq(subscriptions.userId, userId), isNull(subscriptions.unsubscribedAt)));
-
-    const existingUrls = new Set(existingSubscriptions.map((s) => s.feedUrl));
-
-    // Collect all unique tag names from categories in the import
-    const tagNames = new Set<string>();
-    for (const feed of importRecord.feedsData) {
-      if (feed.category) {
-        for (const categoryName of feed.category) {
-          tagNames.add(categoryName);
-        }
-      }
-    }
-
-    // Get or create tags for each unique category name
-    // Store full tag info for subscription_created events
-    const tagNameToInfo = new Map<string, { id: string; name: string; color: string | null }>();
-    if (tagNames.size > 0) {
-      // Get existing tags for this user
-      const existingTags = await db
-        .select({ id: tags.id, name: tags.name, color: tags.color })
-        .from(tags)
-        .where(eq(tags.userId, userId));
-
-      for (const existingTag of existingTags) {
-        if (tagNames.has(existingTag.name)) {
-          tagNameToInfo.set(existingTag.name, {
-            id: existingTag.id,
-            name: existingTag.name,
-            color: existingTag.color,
-          });
-        }
-      }
-
-      // Batch create tags that don't exist (avoids N+1 queries)
-      const now = new Date();
-      const tagsToCreate = Array.from(tagNames)
-        .filter((tagName) => !tagNameToInfo.has(tagName))
-        .map((tagName) => ({
-          id: generateUuidv7(),
-          userId,
-          name: tagName,
-          createdAt: now,
-        }));
-
-      if (tagsToCreate.length > 0) {
-        await db.insert(tags).values(tagsToCreate);
-        for (const tag of tagsToCreate) {
-          tagNameToInfo.set(tag.name, { id: tag.id, name: tag.name, color: null });
-        }
-        logger.debug("OPML import: created tags", {
-          count: tagsToCreate.length,
-          tagNames: tagsToCreate.map((t) => t.name),
-          userId,
-        });
-      }
-    }
-
-    // Process each feed
-    const results: OpmlImportFeedResult[] = [];
-    let imported = 0;
-    let skipped = 0;
-    let failed = 0;
-    let feedsProcessedSinceLastDbUpdate = 0;
-
-    // Helper to update import progress in database (batched to reduce load)
-    const flushProgressToDb = async (force = false) => {
-      feedsProcessedSinceLastDbUpdate++;
-      if (force || feedsProcessedSinceLastDbUpdate >= OPML_IMPORT_DB_BATCH_SIZE) {
-        await db
-          .update(opmlImports)
-          .set({
-            importedCount: imported,
-            skippedCount: skipped,
-            failedCount: failed,
-            results,
-            updatedAt: new Date(),
-          })
-          .where(eq(opmlImports.id, importId));
-        feedsProcessedSinceLastDbUpdate = 0;
-      }
-    };
-
-    for (const opmlFeed of importRecord.feedsData) {
-      const feedUrl = opmlFeed.xmlUrl;
-      const feedTitle = opmlFeed.title ?? null;
-
-      // Check if already subscribed
-      if (existingUrls.has(feedUrl)) {
-        results.push({
-          url: feedUrl,
-          title: feedTitle,
-          status: "skipped",
-          error: "Already subscribed",
-        });
-        skipped++;
-
-        // Publish progress event
-        await publishImportProgress(userId, importId, feedUrl, "skipped", {
-          imported,
-          skipped,
-          failed,
-          total: importRecord.totalFeeds,
-        });
-
-        // Update import record with current progress (batched)
-        await flushProgressToDb();
-
-        continue;
-      }
-
-      try {
-        const subscriptionResult = await createSubscription(db, userId, {
-          url: feedUrl,
-          title: feedTitle,
-          siteUrl: opmlFeed.htmlUrl ?? null,
-        });
-
-        const actualSubscriptionId = subscriptionResult.subscriptionId;
-        const feedId = subscriptionResult.feed.id;
-
-        // Associate subscription with tags from categories
-        if (opmlFeed.category && opmlFeed.category.length > 0) {
-          const tagInfos = opmlFeed.category
-            .map((categoryName) => tagNameToInfo.get(categoryName))
-            .filter(
-              (info): info is { id: string; name: string; color: string | null } =>
-                info !== undefined
-            );
-
-          if (tagInfos.length > 0) {
-            // Delete any existing subscription_tags (in case of reactivation)
-            await db
-              .delete(subscriptionTags)
-              .where(eq(subscriptionTags.subscriptionId, actualSubscriptionId));
-
-            // Insert new subscription_tags entries
-            const tagNow = new Date();
-            await db.insert(subscriptionTags).values(
-              tagInfos.map((tagInfo) => ({
-                subscriptionId: actualSubscriptionId,
-                tagId: tagInfo.id,
-                createdAt: tagNow,
-              }))
-            );
-
-            // Publish update event so the UI picks up the tags
-            publishSubscriptionUpdated(
-              userId,
-              actualSubscriptionId,
-              tagNow,
-              tagInfos,
-              null // no custom title
-            ).catch((err) => {
-              logger.error("Failed to publish subscription_updated event", {
-                err,
-                userId,
-                feedId,
-              });
-            });
-
-            logger.debug("OPML import: associated subscription with tags", {
-              subscriptionId: actualSubscriptionId,
-              tagIds: tagInfos.map((t) => t.id),
-              feedUrl,
-            });
-          }
-        }
-
-        // Add to existing URLs set to prevent duplicates within this import
-        existingUrls.add(feedUrl);
-
-        results.push({
-          url: feedUrl,
-          title: feedTitle,
-          status: "imported",
-          feedId,
-          subscriptionId: actualSubscriptionId,
-        });
-        imported++;
-
-        // Publish progress event
-        await publishImportProgress(userId, importId, feedUrl, "imported", {
-          imported,
-          skipped,
-          failed,
-          total: importRecord.totalFeeds,
-        });
-
-        // Update import record with current progress (batched)
-        await flushProgressToDb();
-
-        logger.info("OPML import: feed imported", { feedUrl, userId, importId });
-      } catch (error) {
-        const errorMessage = error instanceof Error ? error.message : "Unknown error";
-        results.push({
-          url: feedUrl,
-          title: feedTitle,
-          status: "failed",
-          error: errorMessage,
-        });
-        failed++;
-
-        // Publish progress event
-        await publishImportProgress(userId, importId, feedUrl, "failed", {
-          imported,
-          skipped,
-          failed,
-          total: importRecord.totalFeeds,
-        });
-
-        // Update import record with current progress (batched)
-        await flushProgressToDb();
-
-        logger.warn("OPML import: feed import failed", {
-          feedUrl,
-          userId,
-          importId,
-          error: errorMessage,
-        });
-      }
-    }
-
-    // Mark import as completed
-    const completedAt = new Date();
-    await db
-      .update(opmlImports)
-      .set({
-        status: "completed",
-        importedCount: imported,
-        skippedCount: skipped,
-        failedCount: failed,
-        results,
-        completedAt,
-        updatedAt: completedAt,
-      })
-      .where(eq(opmlImports.id, importId));
-
-    // Publish completed event
-    await publishImportCompleted(userId, importId, {
-      imported,
-      skipped,
-      failed,
-      total: importRecord.totalFeeds,
-    });
-
-    logger.info("OPML import completed", {
-      importId,
-      userId,
-      imported,
-      skipped,
-      failed,
-      total: importRecord.totalFeeds,
-    });
-
-    return {
-      success: true,
-      nextRunAt: new Date(Date.now() + 365 * 24 * 60 * 60 * 1000), // Far future - one-time job
-      metadata: {
-        imported,
-        skipped,
-        failed,
-        total: importRecord.totalFeeds,
-      },
-    };
-  } catch (error) {
-    const errorMessage = error instanceof Error ? error.message : "Unknown error";
-
-    // Mark import as failed
-    await db
-      .update(opmlImports)
-      .set({
-        status: "failed",
-        error: errorMessage,
-        updatedAt: new Date(),
-      })
-      .where(eq(opmlImports.id, importId));
-
-    logger.error("OPML import job failed", { importId, error: errorMessage });
-
-    return {
-      success: false,
-      nextRunAt: new Date(Date.now() + 365 * 24 * 60 * 60 * 1000), // Far future - don't retry
-      error: errorMessage,
-    };
-  }
 }
