+++ tests/integration/user-entries-subscription-id.test.ts
@@ -26,7 +26,7 @@ import {
 } from "../../src/server/db/schema";
 import { generateUuidv7 } from "../../src/lib/uuidv7";
 import { createUserEntriesForFeed } from "../../src/server/feed/entry-processor";
-import { migrateSubscriptionsToExistingFeed } from "../../src/server/jobs/handlers";
+import { migrateSubscriptionsToExistingFeed } from "../../src/server/jobs/handlers/fetch-feed";
 import { createSubscription } from "../../src/server/services/subscriptions";
 import { createTestEntry, createTestFeed, createTestSubscription, createTestUser } from "./helpers";
